pyproject.toml
[tool.poetry]
name = "crest-test"
version = "0.1.0"
description = ""
authors = ["KrunalLaheri <laherikrunal5@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.8"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
django-filter = "^25.2"
django-cors-headers = "^4.9.0"
python-decouple = "^3.8"
pillow = "^12.0.0"
openpyxl = "^3.1.5"
django-extensions = "^4.1"
psycopg2-binary = "^2.9.11"


[tool.poetry.group.dev.dependencies]
black = "^25.9.0"
flake8 = "^7.3.0"
isort = "^7.0.0"
pytest = "^8.4.2"
pytest-django = "^4.11.1"
pytest-cov = "^7.0.0"
django-debug-toolbar = "^6.1.0"
ipython = "^9.7.0"

[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_glob = ["*/migrations/*"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
